fix: Remove Girl Develop It from orgs list
In light of recent Girl Develop It controversies about racism and the silencing of its' members, I no longer feel comfortable supporting GDI when I share this wonderful collection of documentation (which is to almost every woman I know). This PR removes them from the main file.

For more information, check out this resource: http://an-open-letter-to-gdi-board.com/